# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Powersport Bike Marketplace API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /requests:
    get:
      operationId: getRequests
      tags: [requests]
      summary: List all user interest requests (admin only)
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: bikeId
          in: query
          schema:
            type: number
      responses:
        "200":
          description: Requests list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BikeRequest"
    post:
      operationId: createRequest
      tags: [requests]
      summary: Submit a bike interest request
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateRequestInput"
      responses:
        "201":
          description: Request created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BikeRequest"
  /requests/{id}:
    get:
      operationId: getRequest
      tags: [requests]
      summary: Get a single request (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "200":
          description: Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BikeRequest"
    patch:
      operationId: updateRequestStatus
      tags: [requests]
      summary: Update request status (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateRequestStatusInput"
      responses:
        "200":
          description: Request updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BikeRequest"
    delete:
      operationId: deleteRequest
      tags: [requests]
      summary: Delete a request (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
components:
  schemas:
...
    BikeRequest:
      type: object
      properties:
        id:
          type: number
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        location:
          type: string
        bikeId:
          type: number
          nullable: true
        bikeName:
          type: string
          nullable: true
        message:
          type: string
          nullable: true
        status:
          type: string
        createdAt:
          type: string
      required:
        - id
        - name
        - email
        - phone
        - location
        - status
        - createdAt
    CreateRequestInput:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        location:
          type: string
        bikeId:
          type: number
          nullable: true
        bikeName:
          type: string
          nullable: true
        message:
          type: string
          nullable: true
      required:
        - name
        - email
        - phone
        - location
    UpdateRequestStatusInput:
      type: object
      properties:
        status:
          type: string
      required:
        - status
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success
...